rename: change header's component name as CamelCase
