+++ components/header/HeaderList.tsx
@@ -3,7 +3,7 @@
 import Link from "next/link";
 import { usePathname } from "next/navigation";
 
-import styles from "./header-list.module.scss";
+import styles from "./HeaderList.module.scss";
 import { checkIsCurrentPath } from "@/utils/header";
 
 export type HeaderListProps = {
