+++ components/header/mobile/MobileOpenButton.module.scss
@@ -1,3 +1,5 @@
+@use "../../../styles/variables.scss" as *;
+
 .wrapper {
   display: none;
   cursor: pointer;
