+++ components/header/mobile/MobileOpenButton.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import styles from "./mobile-open-button.module.scss";
+import styles from "./MobileOpenButton.module.scss";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faBars } from "@fortawesome/free-solid-svg-icons";
